Added an additional right bar button item to sort users based off distance from user's current location.

+++ SLUGGR/Users.swift
@@ -27,7 +27,7 @@ class Users: NSObject {
     var userWorkLat :Float?
     var userWorkLong :Float?
     var userWorkLocale :String?
-    var userDistance :Float?
+    var userDistance :Double?
     var userDistanceType :String?
     
 }
